requirements.txt
django==1.11.5certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.33.0
django-oauth-toolkit==1.0.0
django-rest-auth==0.9.1
django-rest-swagger==2.0.7
djangorestframework==3.5.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.4
openapi-codec==1.3.2
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.9.0
uritemplate==3.0.0
urllib3==1.22
